src/visual/set-up/config/material/shaders/build-shader/fragment-effects/effects/brightness/brightnessTransform.ts
import { BrightnessFragmentEffectProps } from "../../../types";

export const brightnessTransform = (
  fragName: string,
  previousFragName: string,
  brightnessParameters: BrightnessFragmentEffectProps
) => {
  const { declareInTransform } = brightnessParameters;
  const fragmentColorInstantiation = `vec4 ${fragName} = ${previousFragName};`;
  const transformation = `
        // BRIGHTNESS
        ${declareInTransform ? fragmentColorInstantiation : ""}
        vec4 brightnessColor = ${fragName} * uBrightness;
        ${fragName} = ${fragName} * brightnessColor;
      `;
  return { fragmentColorInstantiation, transformation };
};
